requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
plotly
scikit-learn
nltk
pillow
tqdm
